Fixed off by one errors, added better file naming.
Fixed the errors that were causing the state machine to output the
incorrect bits. The error was using 6 bits for state instead of 7, as is
specified in the provided document with project 02. Also added a better
file output naming convention where the original file name is
concatanated with the extension ".hex", the existing extension removed.